Move sparse array fixtures into conftest
This is so that they can be accessed by other tests. We will soon
access them in test_image.
