+++ tests/conftest.py
@@ -1,7 +1,11 @@
+import copy
 import io
 
+import numpy as np
 import pytest
 import requests
+
+from stempy.io.sparse_array import SparseArray
 
 
 DATA_URLS = {
@@ -34,3 +38,39 @@ def electron_data_small():
 @pytest.fixture
 def electron_data_large():
     return io_object('electron_large')
+
+
+# SparseArray fixtures
+
+
+@pytest.fixture
+def sparse_array_small(electron_data_small):
+    kwargs = {
+        'dtype': np.uint64,
+    }
+    array = SparseArray.from_hdf5(electron_data_small, **kwargs)
+
+    # Perform some slicing so we don't blow up CI memory when we
+    # do a full expansion.
+    return array[:40:2, :40:2]
+
+
+cached_full_array_small = None
+
+
+@pytest.fixture
+def full_array_small(sparse_array_small):
+    global cached_full_array_small
+
+    if cached_full_array_small is None:
+        # Don't allow this fixture to modify the other fixture
+        array = copy.deepcopy(sparse_array_small)
+
+        # Have to change these so we won't return a SparseArray,
+        # and allow it to return a fully expanded array
+        array.sparse_slicing = False
+        array.allow_full_expand = True
+
+        cached_full_array_small = array[:]
+
+    return cached_full_array_small
